WIP EvictionStrategy based compaction for MemoryCache
