+++ test/Microsoft.Extensions.Caching.Memory.Tests/CacheEntryScopeExpirationTests.cs
@@ -22,7 +22,6 @@ namespace Microsoft.Extensions.Caching.Memory
             return new MemoryCache(new MemoryCacheOptions()
             {
                 Clock = clock,
-                CompactOnMemoryPressure = false,
             });
         }
