+++ test/Microsoft.Extensions.Caching.Memory.Tests/TimeExpirationTests.cs
@@ -16,7 +16,6 @@ namespace Microsoft.Extensions.Caching.Memory
             return new MemoryCache(new MemoryCacheOptions()
             {
                 Clock = clock,
-                CompactOnMemoryPressure = false,
             });
         }
