+++ test/Microsoft.Extensions.Caching.Memory.Tests/TokenExpirationTests.cs
@@ -23,7 +23,6 @@ namespace Microsoft.Extensions.Caching.Memory
             return new MemoryCache(new MemoryCacheOptions()
             {
                 Clock = clock,
-                CompactOnMemoryPressure = false,
             });
         }
